Added viewModel layer.
