--- app/src/main/java/com/gb/stopwatch/MainActivity.kt
@@ -1,63 +1,0 @@
-package com.gb.stopwatch
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.gb.stopwatch.databinding.ActivityMainBinding
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.SupervisorJob
-import kotlinx.coroutines.flow.collect
-import kotlinx.coroutines.launch
-
-class MainActivity : AppCompatActivity() {
-
-    private val binding: ActivityMainBinding by lazy {
-        ActivityMainBinding.inflate(layoutInflater)
-    }
-
-    private val timestampProvider = object : TimestampProvider {
-        override fun getMilliseconds(): Long {
-            return System.currentTimeMillis()
-        }
-    }
-    private val stopwatchListOrchestrator = StopwatchListOrchestrator(
-        StopwatchStateHolder(
-            StopwatchStateCalculator(
-                timestampProvider,
-                ElapsedTimeCalculator(timestampProvider)
-            ),
-            ElapsedTimeCalculator(timestampProvider),
-            TimestampMillisecondsFormatter()
-        ),
-        CoroutineScope(
-            Dispatchers.Main
-                    + SupervisorJob()
-        )
-    )
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(binding.root)
-
-        with(binding) {
-            CoroutineScope(
-                Dispatchers.Main
-                        + SupervisorJob()
-            ).launch {
-                stopwatchListOrchestrator.ticker.collect {
-                    textTime.text = it
-                }
-            }
-
-            buttonStart.setOnClickListener {
-                stopwatchListOrchestrator.start()
-            }
-            buttonPause.setOnClickListener {
-                stopwatchListOrchestrator.pause()
-            }
-            buttonStop.setOnClickListener {
-                stopwatchListOrchestrator.stop()
-            }
-        }
-    }
-}
