+++ app/src/main/java/com/gb/stopwatch/viewmodel/MainViewModel.kt
@@ -1,0 +1,55 @@
+package com.gb.stopwatch.viewmodel
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import com.gb.stopwatch.*
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.SupervisorJob
+import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.launch
+
+class MainViewModel : ViewModel() {
+
+    private val _liveData: MutableLiveData<String> = MutableLiveData()
+    val liveData: LiveData<String> get() = _liveData
+
+    init {
+        CoroutineScope(
+            Dispatchers.Main
+                    + SupervisorJob()
+        ).launch {
+            stopwatchListOrchestrator.ticker.collect {
+                _liveData.postValue(it)
+            }
+        }
+    }
+
+    private val timestampProvider = object : TimestampProvider {
+        override fun getMilliseconds(): Long {
+            return System.currentTimeMillis()
+        }
+    }
+
+    private val stopwatchListOrchestrator = StopwatchListOrchestrator(
+        StopwatchStateHolder(
+            StopwatchStateCalculator(
+                timestampProvider,
+                ElapsedTimeCalculator(timestampProvider)
+            ),
+            ElapsedTimeCalculator(timestampProvider),
+            TimestampMillisecondsFormatter()
+        ),
+        CoroutineScope(
+            Dispatchers.Main
+                    + SupervisorJob()
+        )
+    )
+
+    fun start() = stopwatchListOrchestrator.start()
+
+    fun pause() = stopwatchListOrchestrator.pause()
+
+    fun stop() = stopwatchListOrchestrator.stop()
+}
